Add warning state to input fields
