+++ src/UI/Pages/Settings/Tabs/Configuration/Components/InputActions.tsx
