+++ src/UI/Pages/Settings/Tabs/Configuration/Components/index.ts
@@ -1,0 +1,1 @@
+export * from "./InputRow";
