+++ src/UI/Pages/Settings/Tabs/Configuration/Container.tsx
@@ -3,7 +3,7 @@ import { Alert, AlertActionCloseButton } from "@patternfly/react-core";
 import { TableComposable, Tbody } from "@patternfly/react-table";
 import styled from "styled-components";
 import { EnvironmentSettings } from "@/Core";
-import { InputRow } from "./InputRow";
+import { InputRow } from "./Components";
 
 interface Props {
   infos: EnvironmentSettings.InputInfo[];
